+++ packages/firebase_bootstrap/lib/src/firebase_bootstrap_base.dart
@@ -2,6 +2,8 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_crashlytics/firebase_crashlytics.dart';
 import 'package:flutter/foundation.dart';
 
+import 'crashlytics_error_filter.dart';
+
 /// 初始化 Firebase
 class FirebaseInitializer {
   FirebaseInitializer._();
@@ -12,32 +14,31 @@ class FirebaseInitializer {
       FirebaseCrashlytics.instance
           .setCrashlyticsCollectionEnabled(enableCrashlytics ?? !kDebugMode);
 
-      const fatalError = true;
-
       FlutterError.onError = (errorDetails) {
+        // 保留 Flutter 的红屏提示
+        // FlutterError.presentError(errorDetails);
+
         if (kDebugMode) {
           FlutterError.dumpErrorToConsole(errorDetails);
-        } else {
-          if (fatalError) {
-            FirebaseCrashlytics.instance.recordFlutterFatalError(errorDetails);
-          } else {
-            FirebaseCrashlytics.instance.recordFlutterError(errorDetails);
-          }
+          return;
         }
+        if (CrashlyticsErrorFilter.shouldIgnoreFlutterError(errorDetails)) {
+          // 日志
+          return;
+        }
+        FirebaseCrashlytics.instance.recordFlutterFatalError(errorDetails);
       };
       PlatformDispatcher.instance.onError = (error, stack) {
+        var errorDetails = FlutterErrorDetails(exception: error, stack: stack);
         if (kDebugMode) {
-          FlutterError.dumpErrorToConsole(
-            FlutterErrorDetails(exception: error, stack: stack),
-          );
-        } else {
-          if (fatalError) {
-            FirebaseCrashlytics.instance.recordError(error, stack, fatal: true);
-          } else {
-            // If you want to record a "non-fatal" exception
-            FirebaseCrashlytics.instance.recordError(error, stack);
-          }
+          FlutterError.dumpErrorToConsole(errorDetails);
+          return true;
         }
+        if (CrashlyticsErrorFilter.shouldIgnoreFlutterError(errorDetails)) {
+          // 日志
+          return true;
+        }
+        FirebaseCrashlytics.instance.recordError(error, stack, fatal: true);
         return true;
       };
     } catch (e) {