+++ packages/firebase_bootstrap/CHANGELOG.md
@@ -4,4 +4,4 @@
 
 ## 1.0.1
 
-- Crashlytics error filter.
+- Filter out the failed image loading due to network anomalies from the fatal reports.